Webpack basic configuration
