--- rough.css
@@ -1,144 +1,0 @@
-body {
-    /* margin-bottom: 0; */
-    font-family: 'Roboto', sans-serif;
-    /* margin: 20 20 0 20; */
-}
-
-
-h3 {
-    color:#16235a;
-    font-weight: bolder;
-    margin: 10;
-}
-
-.flex-container {
-    display: flex;
-    justify-content: flex-start;
-    height: 100vh;
-}
-
-.pane {  
-    flex-basis: 30%;
-    /* change */
-    width: 30%;             
-    background-color: #ebf2ea;
-}
-
-.main {
-    flex-basis: 70%;
-    /* change */
-    width: 70%;
-    background-color: #0abda0;
-}
-
-
-#paneButtons {
-    display: flex;
-    width: 100%; 
-    justify-content: flex-start;
-    flex-wrap: wrap;
-}
-
-#newList { 
-    padding: 2px;
-    margin: 5 10;
-    width: calc(100% - 22px)
-}
-
-
-.paneItems {
-    flex-basis: 90%; 
-    /* width: 90%; */
-    text-decoration: underline;
-    word-wrap: break-word;
-    border-radius: 0%;
-    /* display: block; */
-    /* margin: 5 10; */
-    /* width: calc(100% - 22px) */
-}
-
-
-.paneDelete {
-    flex-basis: 10%;
-    /* width: 10%; */
-    float: right;
-    border-radius: 0%;
-    /* padding: 0;
-    width: 34px;
-    border-radius: 10px; */
-}
-
-
-
-.inputDiv {
-    display: flex;
-    width: 70%;
-    margin: 10px auto;
-}
-
-.inputTask {
-    /* width: calc(90% - 3px); */
-    flex-basis: 90%;;
-    display: initial;
-}
-
-#add {
-    /* width: calc(10% - 8px); */
-    flex-basis: 10%;
-    padding: 3px;
-}
-
-.myList {
-    /* background-color: beige; */
-    /* border: 2px solid green; */
-    width: 70%;
-    margin: auto;
-    padding: 0;
-    list-style-type: none;
-    box-shadow: 0 10px 20px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23);
-    background-color: rgb(247, 241, 189);
-    max-height: 600px;
-    overflow-y: scroll;
-}
-
-.listItem {
-    padding: 21px;
-    margin: 11px 0;
-    /* background-color: #f3e7e7;  */
-    /* border-bottom: black solid 1px; */
-     /* box-shadow: 0 10px 20px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23);    */
-    /* font-style: oblique; */
-    word-wrap: break-word;
-    /* background-color: khaki;                                         */
-}
-
-/* .listItem:hover {
-    padding: 21px;
-    margin: 11px 0;
-    background-color: #f3e7e7; 
-    border-bottom: black solid 1px; 
-    box-shadow: 2px 14px 16px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);
-} */
-
-.delete {
-    /* background-color:rgb(250, 64, 64); */
-    float: right;
-    /* border: none; */
-    /* border-radius: 20px;
-    width: 7%;  */
-    padding: 0;
-    width: 34px;
-    border-radius: 10px;
-}
-
-.check{
-    margin-right: 7px;
-}
-
-.strike {
-    text-decoration: line-through;
-}
-
-
-
-
