--- test.html
@@ -1,13 +1,0 @@
-<html>
-    <head>
-        <title>Test</title>
-       
-        <script src="todoUtils.js"></script>
-        <script src="test.js"></script>
-    </head>
-    <body>
-        <div id="root">
-
-        </div>
-    </body>
-</html>
